Split the Distribution.Hackage.DB module into two parts: the "unparsed" and the "parsed" module.
The parsed Hackage DB is visible by default and it corresponds to the API this
package offered prior to version 1.9, i.e. users of hackage-db won't notice any
difference. The new unparsed module exposed the Hackage tarball as text, i.e.
it gives access to the raw Cabal files. The parsed API is now a thin layer on
top of the unparsed one.